# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tutores": {
      "get": {
        "operationId": "TutorController_listar",
        "summary": "API para obtener el listado de tutores",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      },
      "post": {
        "operationId": "TutorController_crear",
        "summary": "API para crear un nuevo tutor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo tutor utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearTutorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      }
    },
    "/tutores/{id}": {
      "get": {
        "operationId": "TutorController_obtenerPorId",
        "summary": "API para obtener un tutor por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tutor que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      },
      "patch": {
        "operationId": "TutorController_actualizar",
        "summary": "API para actualizar un tutor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tutor que deseas actualizar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un tutor existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarTutorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      }
    },
    "/tutores/eliminar/{id}": {
      "delete": {
        "operationId": "TutorController_eliminar",
        "summary": "API para eliminar un tutor por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tutor que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      }
    },
    "/tutores/soft-delete/{id}": {
      "patch": {
        "operationId": "TutorController_softDelete",
        "summary": "API para eliminar un tutor por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tutor que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      }
    },
    "/tutores/restore/{id}": {
      "patch": {
        "operationId": "TutorController_restore",
        "summary": "API para restaurar un tutor eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del tutor que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tutor"
        ]
      }
    },
...
    "/usuarios": {
      "get": {
        "operationId": "UsuarioController_listar",
        "summary": "API para obtener el listado de usuarios",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      },
      "post": {
        "operationId": "UsuarioController_crear",
        "summary": "API para crear un nuevo usuario",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo usuario utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "operationId": "UsuarioController_obtenerPorId",
        "summary": "API para obtener un usuario por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      },
      "patch": {
        "operationId": "UsuarioController_actualizar",
        "summary": "API para actualizar un usuario",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario que deseas actualizar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un usuario existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      }
    },
    "/usuarios/eliminar/{id}": {
      "delete": {
        "operationId": "UsuarioController_eliminar",
        "summary": "API para eliminar un usuario por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      }
    },
    "/usuarios/soft-delete/{id}": {
      "patch": {
        "operationId": "UsuarioController_softDelete",
        "summary": "API para eliminar un usuario por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      }
    },
    "/usuarios/restore/{id}": {
      "patch": {
        "operationId": "UsuarioController_restore",
        "summary": "API para restaurar un usuario eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del usuario que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Usuario"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend",
    "description": "Documentación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "CrearTutorDto": {
        "type": "object",
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "example": "Arturo Rodrigo Perez Chavez"
          },
          "carnetIdentidad": {
            "type": "string",
            "example": "8526947"
          }
        },
        "required": [
          "nombreCompleto",
          "carnetIdentidad"
        ]
      },
      "ActualizarTutorDto": {
        "type": "object",
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "example": "Arturo Rodrigo Perez Chavez"
          },
          "carnetIdentidad": {
            "type": "string",
            "example": "8526947"
          }
        },
        "required": [
          "nombreCompleto",
          "carnetIdentidad"
        ]
      },
...
      "CrearUsuarioDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user"
          },
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd"
          },
          "rol": {
            "type": "string",
            "enum": [
              "admin",
              "user",
              "guest"
            ],
            "example": "user",
            "default": "user"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "ActualizarUsuarioDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user123"
          },
          "email": {
            "type": "string",
            "example": "example@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd"
          },
          "rol": {
            "type": "string",
            "enum": [
              "admin",
              "user",
              "guest"
            ],
            "example": "user"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "rol"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}